// app/infrastructure/repositories/types/ProfileStatsPersistence.ts
export interface ProfileStatsPersistence {
  userId: string;
  numberOfMatches: string;
  assists: string;
  shotsOnTarget: string;
  shotsOffTarget: string;
  shotsHitPost: string;
  goalFromInsideBox: string;
  goalFromOutsideBox: string;
  ownGoal: string;
  dribbleSuccess: string;
  dribbleFailed: string;
  passCompleted: string;
  passFailed: string;
  tackle: string;
  interception: string;
  block: string;
  possessionLost: string;
  ballRecovery: string;
  foulCommitted: string;
  foulSuffered: string;
  yellowCard: string;
  redCard: string;
  penaltyWon: string;
  penaltyMissed: string;
  penaltyScored: string;
  penaltyConceded: string;
  freeKickScored: string;
  nutmeg: string;
  lob: string;
  save: string;
  saveInsideBox: string;
  saveOutsideBox: string;
  penaltySave: string;
  goalConceded: string;
  createdAt: Date;
  updatedAt: Date;
}